moved SystemDrives class into system_drives.py
